+++ dhafnck_mcp_main/src/fastmcp/task_management/application/use_cases/unassign_agent.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from typing import List, Optional
 from ...domain.repositories.agent_repository import AgentRepository
 from ...domain.exceptions import AgentNotFoundError, ProjectNotFoundError
-from ...infrastructure.repositories.orm.git_branch_repository import ORMGitBranchRepository
+from ...infrastructure.repositories.repository_factory import RepositoryFactory
 
 logger = logging.getLogger(__name__)
 
@@ -50,7 +50,7 @@ class UnassignAgentUseCase:
     
     def __init__(self, agent_repository: AgentRepository):
         self._agent_repository = agent_repository
-        self._git_branch_repository = ORMGitBranchRepository()
+        self._git_branch_repository = RepositoryFactory.get_git_branch_repository()
     
     
     def execute(self, request: UnassignAgentRequest) -> UnassignAgentResponse: